create api doc using postman

+++ src/main/java/com/cerp/springboot/service/StudentService.java
@@ -8,8 +8,8 @@ public interface StudentService {
 	
 	Student saveStudent(Student student);
 	List<Student> getAllStudent();
-	Student getStudentById(int id);
-	Student updateStudent(Student student, int id);
-	void deleteStudent(int id);
+	Student getStudentById(long id);
+	Student updateStudent(Student student, long id);
+	void deleteStudent(long id);
 
 }
